refactor(capabilities): migrate input, trace, window and test_bench to the struct-hosted actor macro
Finishes the struct-hosted actor-macro migration series (ADR-0122
identity/runtime split, ADR-0123 macro-on-the-struct): the four
remaining impl-hosted caps move `#[actor(singleton)]` onto the
identity ZST in mod.rs and relocate the inline `mod runtime {}` shim
plus the `impl NativeActor` (now `#[runtime]`) into a sibling
runtime.rs. input/subscription.rs is consumed into input/mod.rs +
input/runtime.rs, with InputConfig's gate realigned from
`not(target_family = "wasm")` to `feature = "runtime"`. test_bench.rs
promotes from a flat crate-root file to a test_bench/ directory
submodule per ADR-0121 rule 1.

Closes #2474.
